profitability page and form folders restructure
