+++ lib/app/navigation/app_router.dart
@@ -11,7 +11,6 @@ import '../widget/page/default_page.dart';
 import '../widget/page/edit_cost_price_page.dart';
 import '../widget/page/excel_upload_page.dart';
 import '../widget/page/start_new_profitability_page.dart';
-import '../widget/page/logistic_page.dart';
 import '../widget/page/profitability_page.dart';
 
 part 'app_router.gr.dart';
@@ -27,7 +26,6 @@ class AppRouter extends _$AppRouter {
       AutoRoute(page: CostPriceHistoryRoute.page),
       AutoRoute(page: ExcelUploadRoute.page),
       AutoRoute(page: StartNewProfitabilityRoute.page),
-      AutoRoute(page: LogisticRoute.page),
       AutoRoute(page: ProfitabilityRoute.page),
     ];
   }
