+++ lib/app/widget/page/profitability/size/size_input_widget.dart
