+++ lib/domain/profitability/category_selector/category_selector.dart
