+++ lib/domain/profitability/inputs/cost_input.dart
