+++ lib/domain/profitability/inputs/discount_input.dart
